[feat] 로그인 작업중

+++ src/main/java/com/hamhabocca/dallibocca/login/repository/LoginRepository.java
@@ -3,5 +3,5 @@ package com.hamhabocca.dallibocca.login.repository;
 import com.hamhabocca.dallibocca.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoginRepository extends JpaRepository<Member, Integer> {
+public interface LoginRepository extends JpaRepository<Member, Long> {
 }
